fixed project host cache leak.
