+++ src/EditorFeatures/Test/Workspaces/ProjectCacheHostServiceFactoryTests.cs
@@ -19,7 +19,7 @@ namespace Microsoft.CodeAnalysis.Editor.UnitTests.Workspaces
             // Putting cacheService.CreateStrongReference in a using statement
             // creates a temporary local that isn't collected in Debug builds
             // Wrapping it in a lambda allows it to get collected.
-            var cacheService = new ProjectCacheHostServiceFactory.ProjectCacheService(null, int.MaxValue);
+            var cacheService = new ProjectCacheService(null, int.MaxValue);
             var projectId = ProjectId.CreateNewId();
             var owner = new Owner();
             var instance = new ObjectReference();
@@ -106,7 +106,7 @@ namespace Microsoft.CodeAnalysis.Editor.UnitTests.Workspaces
         [Fact]
         public void TestImplicitCacheKeepsObjectAlive1()
         {
-            var cacheService = new ProjectCacheHostServiceFactory.ProjectCacheService(null, int.MaxValue);
+            var cacheService = new ProjectCacheService(null, int.MaxValue);
             var instance = new object();
             var weak = new WeakReference(instance);
             cacheService.CacheObjectIfCachingEnabledForKey(ProjectId.CreateNewId(), (object)null, instance);
@@ -119,7 +119,7 @@ namespace Microsoft.CodeAnalysis.Editor.UnitTests.Workspaces
         [Fact]
         public void TestImplicitCacheMonitoring()
         {
-            var cacheService = new ProjectCacheHostServiceFactory.ProjectCacheService(null, 10, forceCleanup: true);
+            var cacheService = new ProjectCacheService(null, 10, forceCleanup: true);
             var weak = PutObjectInImplicitCache(cacheService);
 
             var timeout = TimeSpan.FromSeconds(10);
@@ -140,7 +140,7 @@ namespace Microsoft.CodeAnalysis.Editor.UnitTests.Workspaces
             GC.KeepAlive(cacheService);
         }
 
-        private static WeakReference PutObjectInImplicitCache(ProjectCacheHostServiceFactory.ProjectCacheService cacheService)
+        private static WeakReference PutObjectInImplicitCache(ProjectCacheService cacheService)
         {
             var instance = new object();
             var weak = new WeakReference(instance);
@@ -165,7 +165,7 @@ namespace Microsoft.CodeAnalysis.Editor.UnitTests.Workspaces
             var instance = new object();
             var weak = new WeakReference(instance);
 
-            var cacheService = new ProjectCacheHostServiceFactory.ProjectCacheService(workspace, int.MaxValue);
+            var cacheService = new ProjectCacheService(workspace, int.MaxValue);
             using (var cache = cacheService.EnableCaching(project2.Id))
             {
                 cacheService.CacheObjectIfCachingEnabledForKey(project1.Id, (object)null, instance);
@@ -186,7 +186,7 @@ namespace Microsoft.CodeAnalysis.Editor.UnitTests.Workspaces
         public void TestEjectFromImplicitCache()
         {
             List<Compilation> compilations = new List<Compilation>();
-            for (int i = 0; i < ProjectCacheHostServiceFactory.ProjectCacheService.ImplicitCacheSize + 1; i++)
+            for (int i = 0; i < ProjectCacheService.ImplicitCacheSize + 1; i++)
             {
                 compilations.Add(CSharpCompilation.Create(i.ToString()));
             }
@@ -194,8 +194,8 @@ namespace Microsoft.CodeAnalysis.Editor.UnitTests.Workspaces
             var weakFirst = new WeakReference(compilations[0]);
             var weakLast = new WeakReference(compilations[compilations.Count - 1]);
 
-            var cache = new ProjectCacheHostServiceFactory.ProjectCacheService(null, int.MaxValue);
-            for (int i = 0; i < ProjectCacheHostServiceFactory.ProjectCacheService.ImplicitCacheSize + 1; i++)
+            var cache = new ProjectCacheService(null, int.MaxValue);
+            for (int i = 0; i < ProjectCacheService.ImplicitCacheSize + 1; i++)
             {
                 cache.CacheObjectIfCachingEnabledForKey(ProjectId.CreateNewId(), (object)null, compilations[i]);
             }
@@ -218,7 +218,7 @@ namespace Microsoft.CodeAnalysis.Editor.UnitTests.Workspaces
             var weak3 = new WeakReference(comp3);
             var weak1 = new WeakReference(comp1);
 
-            var cache = new ProjectCacheHostServiceFactory.ProjectCacheService(null, int.MaxValue);
+            var cache = new ProjectCacheService(null, int.MaxValue);
             var key = ProjectId.CreateNewId();
             var owner = new object();
             cache.CacheObjectIfCachingEnabledForKey(key, owner, comp1);
